This commit fix following issues: 1. Imporve node-tool, add function to query the status of data migration 2. When add/remove node, add a logic to query whther data migration for previous change membership operation is finished. If not, refuse the operation. 3. In the second stage of change membership, use preAddNode/preRemoveNode instead of syncLeader to make sure the order. 4. Fix a bug of install snapshot when data migration. In this case, it's wrong to judge overlap just by the range of plan index because they are not in the same raft group. So add a logic to distinguish snapshot for catch-up and data migration. For catch-up, new tsfile in data migration will renumber the plan indexs. 5. Fix a bug of repeatlly pull slots which may leader to error. 6. Improvement： remove the syncLeader when install snapshot in the stage of data migration because it's unnecessary. 7. When query, it should consider the case that slot is in data migration, which means that it should merge the resource of local data and previous holder data.

+++ cluster/src/main/java/org/apache/iotdb/cluster/utils/nodetool/function/Partition.java
@@ -78,7 +78,7 @@ public class Partition extends NodeToolCmd {
     } else {
       timeRangeMapRaftGroup.forEach(
           (timeRange, raftGroup) -> msgPrintln(String.format("DATA<%s, %d, %d>\t->\t%s", path,
-              (long) timeRange.getKey(0), (long) timeRange.getKey(1),
+              timeRange.getKey(0), timeRange.getKey(1),
               partitionGroupToString(raftGroup))));
     }
   }
